mqtt-gpio: don't fail on bad initial connect()
Don't fail immediately if the program has difficulty connecting to the MQTT
server when initially starting. It's not unusual for the network to need some
time to come up during initial boot. Instead of failing immediately, simply
wait a period and continue trying.

If the connection is lost after the running for a while, the mosquitto client
library code automatically handles reconnecting.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>