mqtt-gpio.c: use short initial connection attempts
It seems like the first attempt to connect to the mqtt server always fails,
but the second succeeds. Try shorter delays between connect attempts on
program startup, then gradually make the delays longer.

My guess is the real culprit is that the service is being called too soon
during init.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>