CMD: add
Up until now, the only association was between mqtt messages and gpio pins.
Expand how this program works to now also accept an association between mqtt
messages and commands. Issuing a subsribed-to "ON" message will run the given
command in the config file; the "OFF" message will terminate the command.

NOTE: currently the command that is run can not have arguments.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>