CMD: add optional 'oneshot'
If a CMD line includes the optional string 'oneline' at the end, then it is
expected that the CMD will run and terminate relatively quickly. In which
case waitpid() is called immediately to cleanup the defunct process.

Otherwise, without the optional 'oneline', the command is expected to run
when the "ON" message is received, and be terminated when the corresponding
"OFF" message occurs.

USE WITH CARE! Specifying the 'oneshot' option on a long-running CMD will
cause the program to not do anything else until the CMD terminates (if it
ever does).

Signed-off-by: Trevor Woerner <twoerner@gmail.com>