refactor: introduce the idea of the ID
When there was no CMD, everything centred around GPIOs, therefore naming
variables as some variant of a GPIO made sense. But when the CMD was
introduced, it forced a reconsideration of the higher-level roles of various
components. The item that links either GPIO or CMD lines with SUB topic is
better described as an ID. Therefore rename variables, update examples and
comments to use this new nomenclature.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>