refactor: add _p for pointers
My custom is to use "_p" for variable names that are pointers. I mostly
did not use "_p" in this code, but did in some cases. Make the code more
uniform by changing all of them to match, which is my custom. This keeps me
from making pointer-related mistakes (which can take some time to figure
out).

Signed-off-by: Trevor Woerner <twoerner@gmail.com>